+++ JogoDaMemoria/interface.hpp
@@ -3,7 +3,7 @@
 #include <allegro5/bitmap.h>
 
 typedef struct {
-	int id;
+	short int id;
 	int x;
 	int y;
 	ALLEGRO_BITMAP* hud;
